pyproject.toml
[tool.poetry]
name = "iou"
version = "0.1.0"
description = ""
authors = ["Alex Wang <alexwang8880@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^6.0"
argh = "^0.26.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
ptipython = "^1.0.1"

[tool.poetry.scripts]
iou = "iou.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
